+++ src/contracts/ChainContract.ts
@@ -6,6 +6,8 @@ import { Logger } from 'winston';
 import Settings from '../types/Settings';
 import Blockchain from '../lib/Blockchain';
 import { ChainBlockData, ChainFCDsData } from '../models/ChainBlockData';
+import { Validator } from '../types/Validator';
+import { ChainStatus } from '../types/ChainStatus';
 
 @injectable()
 class ChainContract {
@@ -33,10 +35,20 @@ class ChainContract {
     return this.setContract(chainAddress);
   }
 
-  setContract = (chainAddress: string): ChainContract => {
-    this.contract = new Contract(chainAddress, ABI.chainAbi, this.blockchain.provider);
-    return this;
-  };
+  async resolveStatus(): Promise<[address: string, status: ChainStatus]> {
+    const chain = await this.resolveContract();
+    return Promise.all([chain.contract.address, chain.contract.getStatus()]);
+  }
+
+  resolveValidators(chainStatus: ChainStatus): Validator[] {
+    return chainStatus.validators.map((address, i) => {
+      return {
+        id: address,
+        location: chainStatus.locations[i],
+        power: chainStatus.powers[i],
+      };
+    });
+  }
 
   async blocksCountOffset(): Promise<number> {
     await this._assertContract();
@@ -47,31 +59,13 @@ class ChainContract {
     return this.contract.address;
   }
 
-  async getBlockHeight(): Promise<BigNumber> {
+  async getBlockId(): Promise<BigNumber> {
     await this._assertContract();
-    return this.contract.getBlockHeight();
+    return this.contract.getBlockId();
   }
 
-  async getBlockVoters(height: number): Promise<string[]> {
-    await this._assertContract();
-    return this.contract.getBlockVoters(height);
-  }
-
-  async resolveBlockVoters(chainAddress: string, height: number): Promise<string[]> {
-    return this.setContract(chainAddress).getBlockVoters(height);
-  }
-
-  async getBlockVotes(blockHeight: number, voter: string): Promise<BigNumber> {
-    await this._assertContract();
-    return this.contract.getBlockVotes(blockHeight, voter);
-  }
-
-  async resolveBlockVotes(chainAddress: string, blockHeight: number, voter: string): Promise<BigNumber> {
-    return this.setContract(chainAddress).getBlockVotes(blockHeight, voter);
-  }
-
-  async resolveBlockData(chainAddress: string, blockHeight: number): Promise<ChainBlockData> {
-    return this.setContract(chainAddress).contract.getBlockData(blockHeight);
+  async resolveBlockData(chainAddress: string, blockId: number): Promise<ChainBlockData> {
+    return this.setContract(chainAddress).contract.blocks(blockId);
   }
 
   async resolveFCDs(chainAddress: string, keys: string[]): Promise<ChainFCDsData> {
@@ -83,6 +77,11 @@ class ChainContract {
       await this.resolveContract();
     }
   }
+
+  private setContract = (chainAddress: string): ChainContract => {
+    this.contract = new Contract(chainAddress, ABI.chainAbi, this.blockchain.provider);
+    return this;
+  };
 }
 
 export default ChainContract;