+++ src/controllers/ProofsController.ts
@@ -15,10 +15,9 @@ class ProofsController {
   }
 
   index = async (request: Request, response: Response): Promise<void> => {
-    const apiKeyVerificationResult = await this.authUtils.verifyApiKeyFromAuthHeader(request.headers.authorization);
+    const apiKeyVerificationResult = await this.authUtils.verifyApiKey(request, response);
 
     if (!apiKeyVerificationResult.apiKey) {
-      response.status(401).send({ error: apiKeyVerificationResult.errorMessage });
       return;
     }
 
@@ -26,12 +25,16 @@ class ProofsController {
       projectId: apiKeyVerificationResult.apiKey.projectId,
     });
 
-    const block = await Block.findOne({ status: BlockStatus.Finalized }).sort({ height: -1 }).limit(1);
+    const last3 = await Block.find({ status: BlockStatus.Finalized }).sort({ blockId: -1 }).limit(3);
+    const block = await Block.findOne({ status: BlockStatus.Finalized }).sort({ blockId: -1 }).limit(1);
+
+    console.log('last3', last3);
+    console.log(block);
 
     if (block) {
       const keys = (request.query.keys || []) as string[];
-      const leaves = await Leaf.find({ blockId: block.id, key: { $in: keys } });
-      response.send({ data: { block: block, keys: keys, leaves: leaves } });
+      const leaves = await Leaf.find({ blockId: block.blockId.toString(), key: { $in: keys } });
+      response.send({ data: { block, keys, leaves } });
     } else {
       response.send({ data: {} });
     }