+++ src/types/BlockFromPegasus.ts
@@ -4,9 +4,14 @@ export interface BlockFromPegasus {
   _id: string;
   timestamp: Date;
   mintedAt: Date;
-  height: number;
+  blockId: number;
   root: string;
   data: Record<string, HexStringWith0x>;
   numericFcdKeys: Array<string>;
   numericFcdValues: Array<number>;
+  anchor: string;
+  votes: Record<string, string>;
+  power: string;
+  staked: string;
+  minter: string;
 }